+++ examples/counter-wagmi/src/components/ToastContainer.tsx
@@ -1,0 +1,19 @@
+import { Toast } from './Toast';
+import { useToast } from '../hooks/useToast';
+
+export function ToastContainer() {
+  const { toasts, removeToast } = useToast();
+
+  return (
+    <div className="fixed bottom-4 right-4 z-50 space-y-2">
+      {toasts.map(toast => (
+        <Toast
+          key={toast.id}
+          message={toast.message}
+          type={toast.type}
+          onClose={() => removeToast(toast.id)}
+        />
+      ))}
+    </div>
+  );
+}
